build: Try to clean up versioning when creating releases.
Cloud Build appears to set SHORT_SHA and various other
substitution variables (but conspiculously not TAG_NAME) to
the tag name when a git tag is passed as the "commit hash"
when manually running a trigger. Use this to get the version
number that's used in the archive filename and the
executable's -version flag. Annoyingly, I don't know how to
test this change without doing another release.

Also hardcode the executable name in the usage string to
prevent Windows from printing an expanded path there.